Create midtermSubmission
This program is based on techniques learned in CS135 and falls in line with the requirements outlined in the Midterm exam: Programming Portion (midtermB). The program counts the vowels in the provided texts and writes an appropriate output based on the test script. The techniques used in this program are based on Chapters 1 to 6 (week 06) of the course material. The program is designed to be efficient and effective in counting vowels in any given text.